requirements.txt
selenium==4.21.0
beautifulsoup4==4.12.3
webdriver_manager==4.0.1
lxml==5.2.2
selenium-wire==5.1.0
python-dotenv==1.0.1
chromedriver-autoinstaller==0.6.4
openai==1.34.0